gfx,tools: Rework the way per-meshlet joints are stored.
Allows us to support a fairly large number of joints per mesh while
still allowing us to ditch a lot of branch-heavy code. The only real
downside is increased LDS usage.

+++ tools/libgltfimport/gltf_import.h
@@ -531,7 +531,7 @@ private:
   GfxMeshletMetadata                      m_metadata;
   std::vector<char>                       m_buffer;
 
-  std::vector<uint32_t>                   m_localJoints;
+  std::vector<uint16_t>                   m_localJoints;
 
   std::vector<GltfVertex> loadVertices(
     const uint32_t*                     indices,
